scraper/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.1
